Return to main controller page after a new entity is created.
As well, fix some errors in the (temporary) viewAll.jsp template.  This view will likely be replaced by a more enhanced search view, but for now it serves as a placeholder to view all entities.

NOTE: For the viewAll JSP to display, one must replace any JSTL references in within the <td> tag with '$', instead of '#'.  '$' cannot be used in generation, due to conflicts with Seam Render.